requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
python-dotenv==1.0.1
huggingface_hub>=0.25.0
httpx>=0.27.0
